# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products/{id}:
    get:
      operationId: getPublicProduct
      tags: [marketplace]
      summary: Get one public marketplace product with its supplier and category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A public product, its category and its public supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetPublicProductResponse"
        "400":
          description: Malformed product id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No such product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unable to load product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /suppliers/{id}:
    get:
      operationId: getPublicSupplier
      tags: [marketplace]
      summary: Get one public supplier with their published products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A public supplier and their published products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetPublicSupplierResponse"
        "400":
          description: Malformed supplier id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No such supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unable to load supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /suppliers/me/quotes:
    get:
      operationId: listMySupplierQuoteRequests
      tags: [marketplace]
      summary: List quote requests sent to the signed-in supplier
      responses:
        "200":
          description: Quote requests sent to the signed-in supplier
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupplierQuoteRequest"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No supplier profile exists for this account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unable to load quote requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /suppliers/{id}/quote:
    post:
      operationId: createSupplierQuoteRequest
      tags: [marketplace]
      summary: Send a quote request to a supplier, optionally about one product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierQuoteRequestBody"
      responses:
        "201":
          description: Quote request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierQuoteRequest"
        "400":
          description: Invalid quote request, product id, or unknown supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No such supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unable to submit quote request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /categories:
    get:
      operationId: listCategories
      tags: [marketplace]
      summary: List marketplace material categories
      responses:
        "200":
          description: Marketplace material categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketplaceCategory"
        "500":
          description: Unable to load categories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products:
    get:
      operationId: listProducts
      tags: [marketplace]
      summary: List marketplace material products
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
            maxLength: 120
          description: Free-text search across product name, description, unit, supplier, category, city and area. Multi-word queries match all words, and small spelling differences are tolerated when no exact match exists.
        - name: category
          in: query
          required: false
          schema:
            type: string
            maxLength: 64
          description: Filter by category id
        - name: city
          in: query
          required: false
          schema:
            type: string
            maxLength: 120
          description: Filter by city
        - name: supplier
          in: query
          required: false
          schema:
            type: string
            maxLength: 64
          description: Filter by supplier id
      responses:
        "200":
          description: Marketplace products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductPreview"
        "400":
          description: Invalid category or supplier id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unable to load products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    PublicProduct:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
          description: Price in PKR. Null means the supplier has not published a price.
        priceLabel:
          type: ["string", "null"]
          description: 'Rendered fallback for a null price: "Price on request".'
        city:
          type: string
        location:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        featured:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        [id, name, description, unit, price, priceLabel, city, location, imageUrl, featured, createdAt]
    PublicCategoryRef:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required: [id, name]
    PublicSupplier:
      type: object
      description: >-
        Public supplier view. `phone` and `whatsapp` are included
        intentionally from Phase 6C so the WhatsApp and Call actions can open
        the supplier's real stored number; they are never invented and the
        action is simply not rendered when the value is absent. `clerkUserId`
        and all other ownership fields remain private.
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        city:
          type: string
        location:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
      required:
        [id, name, description, city, location, logoUrl, phone, whatsapp]
    GetPublicProductResponse:
      type: object
      properties:
        product:
          $ref: "#/components/schemas/PublicProduct"
        category:
          $ref: "#/components/schemas/PublicCategoryRef"
        supplier:
          $ref: "#/components/schemas/PublicSupplier"
      required: [product, category, supplier]
    CreateSupplierQuoteRequestBody:
      type: object
      properties:
        productId:
          type: string
          maxLength: 64
          description: Optional material the enquiry is about. Must be a product belonging to the supplier in the URL.
        name:
          type: string
          minLength: 2
          maxLength: 120
        email:
          type: string
          format: email
          maxLength: 254
        size:
          type: string
          minLength: 2
          maxLength: 120
        details:
          type: string
          minLength: 2
          maxLength: 2000
        imageUrls:
          type: array
          maxItems: 5
          items:
            type: string
            maxLength: 2048
          description: 'Optional absolute http or https image URLs, up to five. Non-http(s) schemes such as javascript and data URIs are rejected.'
      required: [name, email, size, details]
    SupplierQuoteRequest:
      type: object
      properties:
        id:
          type: string
        productId:
          type: ["string", "null"]
        name:
          type: string
        email:
          type: string
        size:
          type: string
        details:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required: [id, productId, name, email, size, details, imageUrls, createdAt]
    GetPublicSupplierResponse:
      type: object
      properties:
        supplier:
          $ref: "#/components/schemas/PublicSupplier"
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProductPreview"
      required: [supplier, products]
    MarketplaceCategory:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
      required: [id, name, description]
    ProductPreview:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
          description: Price in PKR. Null means the supplier has not published a price.
        priceLabel:
          type: ["string", "null"]
          description: 'Rendered fallback for a null price: "Price on request".'
        category:
          type: string
        categoryId:
          type: string
        supplier:
          type: string
        supplierId:
          type: string
        city:
          type: string
        location:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        featured:
          type: boolean
      required:
        [
          id,
          name,
          description,
          unit,
          price,
          priceLabel,
          category,
          categoryId,
          supplier,
          supplierId,
          city,
          location,
          imageUrl,
          featured,
        ]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...