# lib/api-zod/src/generated/types/getPublicProductResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PublicCategoryRef } from './publicCategoryRef';
import type { PublicProduct } from './publicProduct';
import type { PublicSupplier } from './publicSupplier';

export interface GetPublicProductResponse {
  product: PublicProduct;
  category: PublicCategoryRef;
  supplier: PublicSupplier;
}