# lib/api-zod/src/generated/types/publicProduct.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PublicProduct {
  id: string;
  name: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  unit: string | null;
  /**
     * Price in PKR. Null means the supplier has not published a price.
     * @nullable
     */
  price: number | null;
  /**
     * Rendered fallback for a null price: "Price on request".
     * @nullable
     */
  priceLabel: string | null;
  city: string;
  /** @nullable */
  location: string | null;
  /** @nullable */
  imageUrl: string | null;
  featured: boolean;
  createdAt: Date;
}